requirements.txt
argparse~=1.4.0
peewee~=3.13.3
python-dotenv~=0.14.0
requests~=2.24.0
spotipy~=2.16.0
click~=7.1.2
fuzzywuzzy~=0.18.0
pydub~=0.24.1